# user-api-contract/src/main/resources/v1/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Intuit Demo
servers:
  - url: http://localhost:8081/v1
paths:
  /users:
    get:
      summary: List all users
      operationId: listUsers
      tags:
        - users
      responses:
        '200':
          description: An array of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: Creates a newuser
      operationId: createUser
      tags:
        - users
      requestBody:
        description: User to Create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/current:
    get:
      summary: Info for current user
      operationId: getCurrentUser
      tags:
        - users
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{userId}/groups:
    post:
      description: Sets the groups for a user
      operationId: setGroups
      tags:
        - users
      parameters:
        - name: userId
          in: path
          description: ID of user to update groups
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Groups to set
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Groups'
      responses:
        '200':
          description: user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
    NewUser:
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User"
    Group:
      properties:
        name:
          type: string
    Groups:
      type: array
      items:
        $ref: "#/components/schemas/Group"
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string